requirements.txt
openai-whisper
pyaudio
nltk
numpy

